store/tikv: move some code
