+++ store/tikv/coprocessor.go
@@ -47,6 +47,12 @@ import (
 
 var coprCacheHistogramEvict = tidbmetrics.DistSQLCoprCacheHistogram.WithLabelValues("evict")
 
+// Maximum total sleep time(in ms) for kv/cop commands.
+const (
+	copBuildTaskMaxBackoff = 5000
+	copNextMaxBackoff      = 20000
+)
+
 // CopClient is coprocessor client.
 type CopClient struct {
 	kv.RequestTypeSupportedChecker
@@ -719,69 +725,6 @@ func (worker *copIteratorWorker) handleTaskOnce(bo *Backoffer, task *copTask, ch
 
 	// Handles the response for non-streaming copTask.
 	return worker.handleCopResponse(bo, rpcCtx, &copResponse{pbResp: resp.Resp.(*coprocessor.Response)}, cacheKey, cacheValue, task, ch, nil, costTime)
-}
-
-// ClientHelper wraps LockResolver and RegionRequestSender.
-// It's introduced to support the new lock resolving pattern in the large transaction.
-// In the large transaction protocol, sending requests and resolving locks are
-// context-dependent. For example, when a send request meets a secondary lock, we'll
-// call ResolveLock, and if the lock belongs to a large transaction, we may retry
-// the request. If there is no context information about the resolved locks, we'll
-// meet the secondary lock again and run into a deadloop.
-type ClientHelper struct {
-	lockResolver  *LockResolver
-	regionCache   *RegionCache
-	resolvedLocks *util.TSSet
-	client        Client
-	resolveLite   bool
-	RegionRequestRuntimeStats
-}
-
-// NewClientHelper creates a helper instance.
-func NewClientHelper(store *KVStore, resolvedLocks *util.TSSet) *ClientHelper {
-	return &ClientHelper{
-		lockResolver:  store.GetLockResolver(),
-		regionCache:   store.GetRegionCache(),
-		resolvedLocks: resolvedLocks,
-		client:        store.GetTiKVClient(),
-	}
-}
-
-// ResolveLocks wraps the ResolveLocks function and store the resolved result.
-func (ch *ClientHelper) ResolveLocks(bo *Backoffer, callerStartTS uint64, locks []*Lock) (int64, error) {
-	var err error
-	var resolvedLocks []uint64
-	var msBeforeTxnExpired int64
-	if ch.Stats != nil {
-		defer func(start time.Time) {
-			RecordRegionRequestRuntimeStats(ch.Stats, tikvrpc.CmdResolveLock, time.Since(start))
-		}(time.Now())
-	}
-	if ch.resolveLite {
-		msBeforeTxnExpired, resolvedLocks, err = ch.lockResolver.ResolveLocksLite(bo, callerStartTS, locks)
-	} else {
-		msBeforeTxnExpired, resolvedLocks, err = ch.lockResolver.ResolveLocks(bo, callerStartTS, locks)
-	}
-	if err != nil {
-		return msBeforeTxnExpired, err
-	}
-	if len(resolvedLocks) > 0 {
-		ch.resolvedLocks.Put(resolvedLocks...)
-		return 0, nil
-	}
-	return msBeforeTxnExpired, nil
-}
-
-// SendReqCtx wraps the SendReqCtx function and use the resolved lock result in the kvrpcpb.Context.
-func (ch *ClientHelper) SendReqCtx(bo *Backoffer, req *tikvrpc.Request, regionID RegionVerID, timeout time.Duration, sType kv.StoreType, directStoreAddr string) (*tikvrpc.Response, *RPCContext, string, error) {
-	sender := NewRegionRequestSender(ch.regionCache, ch.client)
-	if len(directStoreAddr) > 0 {
-		sender.SetStoreAddr(directStoreAddr)
-	}
-	sender.Stats = ch.Stats
-	req.Context.ResolvedLocks = ch.resolvedLocks.GetAll()
-	resp, ctx, err := sender.SendReqCtx(bo, req, regionID, timeout, sType)
-	return resp, ctx, sender.GetStoreAddr(), err
 }
 
 const (
